refactor: simplify rate limiter, remove dead code, fix Config wiring
Rate limiter cleanup (-55 lines):
- Delete CheckIP, RecordSuccess, Record (superseded by CheckAndIncrement)
- Promote Lua scripts to package-level vars (avoid per-call SHA1 + alloc)
- Migrate vanity.go and salt.go to atomic CheckAndIncrement (fix TOCTOU)
- Add dedicated batch_agent_info bucket (30/hr, was sharing compute_salt)

Dead code removal:
- Delete recordSuccessfulRelay no-op method + 7 call sites in relay.go
- Remove trailing Record() call in salt.go UploadSalts

Bug fix:
- Wire Config to RouterParams in main.go (TRUST_PROXY env var had no effect)

Minor:
- Use TRUNCATE instead of DELETE for rollback tables (O(1) vs O(n))
- Clean up redundant int256 cast in SubnetManagerUni.unlockCallback

+++ contracts/src/subnets/SubnetManagerUni.sol
@@ -208,8 +208,8 @@ contract SubnetManagerUni is SubnetManager {
         // For !zeroForOne: delta0 > 0 (we receive token0), delta1 < 0 (we owe token1)
         address inputCurrency = zeroForOne ? pk.currency0 : pk.currency1;
         address outputCurrency = zeroForOne ? pk.currency1 : pk.currency0;
-        uint256 inputAmt = uint256(int256(zeroForOne ? -delta0 : -delta1));
-        uint256 outputAmt = uint256(int256(zeroForOne ? delta1 : delta0));
+        uint256 inputAmt = uint256(uint128(zeroForOne ? -delta0 : -delta1));
+        uint256 outputAmt = uint256(uint128(zeroForOne ? delta1 : delta0));
 
         // Slippage check
         if (outputAmt < minOut) revert SlippageExceeded();
